app/requirements.txt
stomp.py==5.0.1
